# docs/openapi_1.0.yaml
openapi: 3.0.3
info:
  title: 广告设计公司客户对接平台 API
  version: "1.0.0"
  description: 1.0 后端接口定义，可导入 Swagger UI 或 YApi。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /api/projects:
    get:
      tags: [Project]
      summary: 当前用户项目列表
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: currentStage
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 项目分页
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProjectPage"
    post:
      tags: [Project]
      summary: 新增项目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "200":
          description: 项目详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProjectVO"
  /api/projects/{id}:
    get:
      tags: [Project]
      summary: 项目详情
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 项目详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProjectVO"
    put:
      tags: [Project]
      summary: 更新项目
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: 项目详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProjectVO"
    delete:
      tags: [Project]
      summary: 删除或取消项目
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/BooleanResult"
  /api/projects/{projectId}/stages:
    get:
      tags: [Stage]
      summary: 项目阶段列表
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: 项目阶段列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProjectStageVO"
  /api/projects/{projectId}/stages/{stageCode}/actions:
    post:
      tags: [Stage]
      summary: 发起阶段确认
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: stageCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStageActionRequest"
      responses:
        "200":
          description: 阶段动作详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StageActionVO"
  /api/stage-actions/{actionId}/confirm:
    post:
      tags: [Stage]
      summary: 确认阶段动作
      parameters:
        - $ref: "#/components/parameters/ActionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageActionResponseRequest"
      responses:
        "200":
          description: 阶段动作详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StageActionVO"
  /api/stage-actions/{actionId}/reject:
    post:
      tags: [Stage]
      summary: 驳回阶段动作
      parameters:
        - $ref: "#/components/parameters/ActionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageActionResponseRequest"
      responses:
        "200":
          description: 阶段动作详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StageActionVO"
  /api/projects/{projectId}/stage-actions:
    get:
      tags: [Stage]
      summary: 阶段动作列表
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: stageCode
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 阶段动作列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/StageActionVO"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BooleanResult:
      description: 布尔结果
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Result"
              - type: object
                properties:
                  data:
                    type: boolean
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    ActionId:
      name: actionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    Size:
      name: size
      in: query
      schema:
        type: integer
        default: 10
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
...
  schemas:
    Result:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
    PageResult:
      type: object
      properties:
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer
...
    CreateProjectRequest:
      type: object
      required: [name, customerId, designerId]
      properties:
        name:
          type: string
        customerId:
          type: integer
          format: int64
        designerId:
          type: integer
          format: int64
        description:
          type: string
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        designerId:
          type: integer
          format: int64
        description:
          type: string
        status:
          type: string
          enum: [IN_PROGRESS, COMPLETED, PAUSED, CANCELLED]
    ProjectVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        customerId:
          type: integer
          format: int64
        customerName:
          type: string
        designerId:
          type: integer
          format: int64
        designerName:
          type: string
        currentStage:
          type: string
        currentStageName:
          type: string
        status:
          type: string
        progress:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ProjectPage:
      allOf:
        - $ref: "#/components/schemas/PageResult"
        - type: object
          properties:
            records:
              type: array
              items:
                $ref: "#/components/schemas/ProjectVO"
    ProjectStageVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        stageCode:
          type: string
        stageName:
          type: string
        sortOrder:
          type: integer
        status:
          type: string
          enum: [TODO, PENDING_CONFIRM, REACHED, REJECTED]
        reachedAt:
          type: string
    CreateStageActionRequest:
      type: object
      properties:
        requestNote:
          type: string
    StageActionResponseRequest:
      type: object
      properties:
        responseNote:
          type: string
    StageActionVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        projectStageId:
          type: integer
          format: int64
        stageCode:
          type: string
        initiatorId:
          type: integer
          format: int64
        initiatorRole:
          type: string
          enum: [CUSTOMER, DESIGNER]
        confirmUserId:
          type: integer
          format: int64
        status:
          type: string
          enum: [PENDING, CONFIRMED, REJECTED, CANCELLED]
        requestNote:
          type: string
        responseNote:
          type: string
        requestedAt:
          type: string
        respondedAt:
          type: string
...